spin_common::paths

Function find_manifest_file_path

pub fn find_manifest_file_path(
    provided_path: Option<impl AsRef<Path>>,
) -> Result<(PathBuf, usize)>

Attempts to find a manifest. If a path is provided, that path is resolved using resolve_manifest_file_path; otherwise, a directory search is carried out using search_upwards_for_manifest. If we had to search, and a manifest is found, a (non-zero) usize is returned indicating how far above the current directory it was found. (A usize of 0 indicates that the manifest was provided or found in the current directory.) This can be used to notify the user that a non-default manifest is being used.