spin_common::paths
Function search_upwards_for_manifest

pub fn search_upwards_for_manifest() -> Option<(PathBuf, usize)>

Starting from the current directory, searches upward through the directory tree for a manifest (that is, a file with the default manifest name spin.toml). If found, the path to the manifest is returned, with a usize indicating how far above the current directory it was found. (A usize of 0 indicates that the manifest was provided or found in the current directory.) This can be used to notify the user that a non-default manifest is being used. If no matching file is found, the function returns None.

The search is abandoned if it reaches the root directory, or the root of a Git repository, without finding a ‘spin.toml’.